Crate database

Re-exports

pub use crate::error::DatabaseError;

pub use crate::rw_pool::write_tx;

Modules

asset

compact_mediafile

episode

error

genre

library

media

mediafile

movie

progress

query_ext

rw_pool

season

tv

user

utils

Macros

opt_update

Functions

generate_key

get_conn

Function which returns a Result<T, E> where T is a new connection session or E is a connection error.

get_conn_logged

Function which returns a Result<T, E> where T is a new connection session or E is a connection error. It takes in a logger instance.

get_conn_memory

set_key

try_get_conn

Type Definitions

DbConnection

Transaction