Module database::cli::args

Structs

CreateIndividualReport

Create Individual Report Struct that takes in arguments to Create

CreateRequirement

Create Requirement Struct that takes in arguments to Create

CreateSprint

Create Sprint Struct that takes in arguments to Create

CreateTeamActivity

Create Team Activity Struct that takes in arguments to Create

CreateTeamReport

Create Team Report Struct that takes in arguments to Create

CreateUser

Create User Struct that takes in arguments to Create

DatabaseArgs

IndividualReportCommand

gets the RequirementCommand to make a sub command with it

RequirementCommand

gets the RequirementCommand to make a sub command with it

SprintCommand

gets the SprintCommand to make a sub command with it

TeamActivityCommand

gets the TeamActivityCommand to make a sub command with it

TeamReportCommand

gets the RequirementCommand to make a sub command with it

UpdateIndividualReport

Update Individual Report Struct that takes in arguments to Update

UpdateRequirement

Update Requirement Struct that takes in arguments to Update

UpdateTeamActivity

Update Team Activity Struct that takes in arguments to Update

UpdateTeamReport

Update Team Report Struct that takes in arguments to Update

UpdateUser

Update User Struct that takes in arguments to Update

UserCommand

gets the UserCommand to make a sub command with it

Enums

EntityType

these are the ENUMS for the commands to be created

IndividualReportSubcommand

splits up what Individual Report Command you want

RequirementSubcommand

splits up what Requirement Command you want

SprintSubcommand

splits up Sprint Command you want

TeamActivitySubcommand

splits up what Team Activity Command you want

TeamReportSubcommand

splits up what Team Report Command you want

UserSubcommand

splits up what UserCommand you want