Trait database::query_ext::QueryExt

pub trait QueryExt<'a, DB: Database> {
    fn bind_all<Vs, V>(self, values: Vs) -> Self
    where
        V: Send + Encode<'a, DB> + Type<DB> + 'a,
        Vs: IntoIterator<Item = V>;
}

Trait contains some extensions for sqlx.

Required Methods

fn bind_all<Vs, V>(self, values: Vs) -> Self where
V: Send + Encode<'a, DB> + Type<DB> + 'a,
Vs: IntoIterator<Item = V>,

Method which allows you to bind several values in one go. This method will accept any container which can be turned into an Iterator of values.