Trait database::Execute

pub trait Execute<'q, DB>: Send where
    DB: Database, {
    pub fn sql(&self) -> &'q str;

    pub fn statement(&self) -> Option<&<DB as HasStatement<'q>>::Statement>;

    pub fn take_arguments(
        &mut self
    ) -> Option<<DB as HasArguments<'q>>::Arguments>;

    pub fn persistent(&self) -> bool;
}

A type that may be executed against a database connection.

Implemented for the following:

&str

Query

Required methods

pub fn sql(&self) -> &'q str

Gets the SQL that will be executed.

pub fn statement(&self) -> Option<&<DB as HasStatement<'q>>::Statement>

Gets the previously cached statement, if available.

pub fn take_arguments(&mut self) -> Option<<DB as HasArguments<'q>>::Arguments>

Returns the arguments to be bound against the query string.

Returning None for Arguments indicates to use a "simple" query protocol and to not prepare the query. Returning Some(Default::default()) is an empty arguments object that will be prepared (and cached) before execution.

pub fn persistent(&self) -> bool

Returns true if the statement should be cached.

Loading content...