Trait database::HasStatement

pub trait HasStatement<'q> {
    type Database: Database;
    type Statement: Statement<'q>;
}

Associate Database with a Statement of a generic lifetime.

The upcoming Rust feature, Generic Associated Types, should obviate the need for this trait.

Associated Types

type Database: Database

type Statement: Statement<'q>

The concrete Statement implementation for this database.

Loading content...